Nog meer fixes

+++ Planets/PlanetsLauncher.cs
@@ -31,7 +31,7 @@ namespace Planets
         {
             var pf = new Playfield();
             // Set player
-            Player p = new Player(new Vector(pf.Size.Width / 2, pf.Size.Height / 2), new Vector(), 1000);
+            Player p = new Player(new Vector((double)pf.Size.Width / 2, (double)pf.Size.Height / 2), new Vector(), 1000);
             pf.CurrentPlayer = p;
 
             // Generate objects
